+++ src/components/tile-select/index.ts
@@ -1,0 +1,7 @@
+export { default as TileSelect } from "./tile-select.component";
+export type {
+  TileSelectProps,
+  TileSelectDeselectEvent,
+} from "./tile-select.component";
+export { default as TileSelectGroup } from "./tile-select-group.component";
+export type { TileSelectGroupProps } from "./tile-select-group.component";